split up extract & data utils
