+++ spiketools/utils/__init__.py
@@ -1,6 +1,6 @@
 """Utility functions."""
 
 from .utils import set_random_seed
-from .data import restrict_range, get_value_by_time, get_value_by_time_range
+from .extract import restrict_range, get_value_by_time, get_value_by_time_range
 from .trials import (epoch_spikes_by_event, epoch_spikes_by_range,
                      epoch_data_by_event, epoch_data_by_range)
